Task_1_1_3

+++ Task_1_1_3/src/main/java/ru/nsu/demidov/expressions/Number.java
@@ -14,7 +14,7 @@ public class Number extends Expression {
      * Number constructor.
      */
 
-    Number(int value) {
+    Number(double value) {
         this.value = value;
     }
